+++ app/client/src/workers/common/DataTreeEvaluator/index.ts
@@ -118,9 +118,6 @@ import { errorModifier } from "workers/Evaluation/errorModifier";
 import JSObjectCollection from "workers/Evaluation/JSObject/Collection";
 import userLogs from "workers/Evaluation/fns/overrides/console";
 import ExecutionMetaData from "workers/Evaluation/fns/utils/ExecutionMetaData";
-import { evalTreeWithChanges } from "workers/Evaluation/evalTreeWithChanges";
-import { WorkerMessenger } from "workers/Evaluation/fns/utils/Messenger";
-import { MAIN_THREAD_ACTION } from "ce/workers/Evaluation/evalWorkerActions";
 
 type SortedDependencies = Array<string>;
 export type EvalProps = {
@@ -1677,58 +1674,6 @@ export default class DataTreeEvaluator {
           },
         },
       );
-    });
-  }
-
-  applySetterMethod(path: string, value: unknown) {
-    const { entityName, propertyPath } = getEntityNameAndPropertyPath(path);
-    const entity = get(this.evalTree, entityName);
-    const updatedProperties: string[][] = [];
-    const overriddenProperties: string[] = [];
-
-    const evalMetaUpdates: EvalMetaUpdates = [];
-
-    if (isWidget(entity)) {
-      overrideWidgetProperties({
-        entity,
-        propertyPath,
-        value,
-        currentTree: this.evalTree,
-        configTree: this.oldConfigTree,
-        evalMetaUpdates,
-        fullPropertyPath: path,
-        isNewWidget: false,
-        shouldUpdateGlobalContext: true,
-        overriddenProperties,
-      });
-
-      overriddenProperties.forEach((propPath) => {
-        updatedProperties.push([entityName, propPath]);
-
-        if (propPath.split(".")[0] === "meta") {
-          const metaPropertyPath = propPath.split(".").slice(1);
-
-          evalMetaUpdates.push({
-            widgetId: entity.widgetId,
-            metaPropertyPath,
-            value,
-          });
-
-          WorkerMessenger.request({
-            method: MAIN_THREAD_ACTION.SET_META_PROP_FROM_SETTER,
-            data: { evalMetaUpdates },
-          });
-        }
-      });
-    }
-
-    set(this.evalTree, path, value);
-    set(self, path, value);
-
-    return new Promise((resolve) => {
-      updatedProperties.push([entityName, propertyPath]);
-
-      evalTreeWithChanges(updatedProperties, true, resolve);
     });
   }
